[SVG] Support transform on SVGSVGElement
This changes adds support for the `transform` property on the
`svg` element. This behavior is in the SVG2 spec and is already
implemented by Firefox.

This is accomplished by incorporating the local transform
(via TransformHelper::ComputeTransformIncludingMotion) into the
existing `local_to_parent_transform_`, and multiplying by the
viewBox transform.

This feature is being tracked in:
https://chromestatus.com/feature/6129368025530368

Bug: 40313130
Change-Id: I8fec7a7e4fd84769ef33f8d7a43aece6daafc56f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6435957
Reviewed-by: Vinay Singh <vinaysingh@microsoft.com>
Commit-Queue: Divyansh Mangal <dmangal@microsoft.com>
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1447198}
NOKEYCHECK=True
GitOrigin-RevId: fc7fe3f8c697adee1f13a48b962b231b9200add4