Migrate link_header_util::ParseLinkHeaderValue() to using string_views.
It was taking string iterators. This is largely an effort to switch it
to using a different HttpUtil::TrimLWS() method, so the old one can
be removed.

This CL does not modify SplitLinkHeader(), which continues to return a
vector of string::iterator std::pairs, though it does add a
ParseLinkHeaderValue() that takes such  values as input, to simplify
migration down the line.

Bug: None
Change-Id: I90536e088ad922568d46aa525ccba9565f9ed9a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6443521
Commit-Queue: mmenke <mmenke@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Slobodan Pejic <slobodan@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1447348}
NOKEYCHECK=True
GitOrigin-RevId: 6a46726d48ae6c438534d64b22015b3ce99d23a6