[Autofill] Make FormControlTypeForAutofill() less sticky
FormControlTypeForAutofill() makes type=password be sticky:
When an <input type=password> changes to <input type=text>, the
FormControlTypeForAutofill() remains kInputPassword.

This CL limits the stickiness to text-type types:
When an <input type=password> changes to <input type=checkbox>, the
FormControlTypeForAutofill() changes to kInputCheckbox.

The motivation is that the status quo leads to surprising code with
unclear behaviour when handling <input type={checkbox,radio}> as in
https://source.chromium.org/chromium/chromium/src/+/main:components/autofill/content/renderer/form_autofill_util.cc;l=285-292;drc=84fae4d9f073d69ce6df1f893104c5cf3dff1f45
and
https://chromium-review.googlesource.com/c/chromium/src/+/6438923/1/components/autofill/content/renderer/autofill_agent.cc#386

Bug: 40554728
Change-Id: I5032ef31047b5b007fcfdfc410b4e2b79728efd0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6443369
Reviewed-by: Dominic Battré <battre@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Christoph Schwering <schwering@google.com>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Reviewed-by: Maria Kazinova <kazinova@google.com>
Cr-Commit-Position: refs/heads/main@{#1447471}
NOKEYCHECK=True
GitOrigin-RevId: f19f7192cdfd45800693ee7ee44c07bb7af787e7