WebGL LRUCanvasResourceProviderCache: Allow BGRA to match RGBA
The LRUCanvasResourceProviderCache maintains a cache of
CanvasResourceProviders. The cache key for these is the size, format,
alpha, and color space.

Sometimes when creating a CanvasResourceProvider (deep in the bowels),
we will use a different format than the requested format. In particular
the caller will request RGBA, but we'll return BGRA. This makes us never
hit the cache (but we will chew up new entries, so we pay the cost and
don't get the benefit).

Record the requested format, and allow a hit if the requested formats
match (or if the CanvasResourceProvider matches).

Ideally all of this would go away and all resource pooling would be done
on the GPU side.

Bug: 407564084 (incidental discovery)
Change-Id: Ied0c674eeb005911d61a021e769d13c9fe998962
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6433208
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1447532}
NOKEYCHECK=True
GitOrigin-RevId: c218105f38f846ca6a36c3cdd42947dfa47f316f