[blink] Return SyncToken from CopyToPlatformInternal()
As a step towards of the full introduction of scoped raster access,
this CL lets CopyToPlatformInternal() return a SyncToken, which in
turn is returned to CopyToPlatformMailbox(). This refactoring is a
prerequisite for follow-up changes in CopyToPlatformMailbox() which
involves waiting on the return SyncToken.

Bug: 40286368
Change-Id: Ifc73ca5c32f9b67ba0a8eb6f4ca932c90535f46a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6454133
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Mingjing Zhang <mjzhang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1447768}
NOKEYCHECK=True
GitOrigin-RevId: d3cada86b02202afc6c9cfa33d3408dfe6894016