[rust] Presubmit: gnrt-auto-generated files shouldn't be manually edited
This CL adds a presubmit based on an approach from
`gpu/vulkan/PRESUBMIT.py` where we check if a CL changes 1) inputs of
code generator (e.g. `gnrt_config.toml` in case of `gnrt`) and 2)
outputs of code generator (e.g. `README.chromium` in case of `gnrt`) and
if only outputs are modified, then reports a presubmit error.

This kind of presubmit can work without actually building and running
the code generator, which is tricky for `gnrt` because building it
depends on (locked) third-party libraries from crates.io and running
it requires connectivity to crates.io (to vendor crates if needed).

Bug: 399934630
Change-Id: I5ea251c5e91f27bf73f246c0f59cb30d617011ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6459941
Reviewed-by: Liza Burakova <liza@chromium.org>
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1448558}
NOKEYCHECK=True
GitOrigin-RevId: 4d4076ee8c1170d3ffe4bd6aed4c80014e89126b