Let MediaElement continue to play after a MediaElementSourceNode is attached
Creating the MediaElementSourceNode causes playback to pause until the node
is connected to a destination audio node.

The proposed fix is to delay the Stop playback of the media element until
the source node is actually connected to a destination.

Bug analysis and solution:
https://docs.google.com/document/d/1gsaVFJ6PvwxGa71jE2bGWtZAQXeZixs0rmUxvsXsGTo/edit?usp=sharing

Bug: 41450896
Change-Id: I95e317e3cbf0046cda5461a2b6612efe6f14b825
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6444424
Commit-Queue: Sunggook Chue <sunggch@microsoft.com>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Michael Wilson <mjwilson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1448560}
NOKEYCHECK=True
GitOrigin-RevId: 610196b281d2ea861996de949f80e47cb2c28cea