Reland [break-iterator] Keep an instance of grapheme cluster iterator
This is a reland of crrev.com/c/6432910, but not created as a
reland due to merge conflicts. The root cause of the revert
was fixed at crbug.com/411169296.

The original message follows.

Constructing `NonSharedCharacterBreakIterator` isn't too
expensive, it doesn't clone the input string, etc.

But it turned out that there are some cases where the
`LazyLineBreakIterator::NextBreakablePositionBreakCharacter`
is very hot, such as computing min-size of 16-bits text with
`word-break: break-word`. In such cases, setting up `utext`,
which involves setting up several function pointers in a
vtable-like data structure, is not trivial.

For example, in the `Editor-CodeMirror` of the Speedometer3,
the `NonSharedCharacterBreakIterator` constructor consumes
0.18%. Keeping an instance in the `LazyLineBreakIterator`
reduces it to 0.06%.

This difference isn't observable in pinpoint. The score vary
from +0.1% to -0.2%; slightly positive if we take an average
of multiple runs.

Bug: 408449458
Change-Id: I9c509ac3db43b9650a715c9926d82c01a7163cab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6472294
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1448815}
NOKEYCHECK=True
GitOrigin-RevId: 9697e644ec5005c9c569a442d96cacde57ca37cf