[perf] Avoid pointers in `GlyphDataRange`
Instead, `GlyphDataRange` stores the pointer to
`ShapeResultRun` and the glyph index.

This change is needed from two perspectives:
1. Required when we move the storages to Oilpan, currently WIP
   at crrev.com/c/6390296
2. To make it safer (no native pointers).

Performance-wise, this change is negative, by ~0.1% for the
Speedometer3 score, and several layout microbenchmarks. But
when combined with the Oilpan-ize, it's ~0.1% positive. The
layout microbenchmarks are mixed though.

From the safety point of view, this CL makes the internal of
`GlyphDataRange` safe, though this CL doesn't try to eliminate
old unsafe public APIs. Replacing usages of the unsafe APIs
with safe `span` APIs will be future work.

Bug: 406127861
Change-Id: I2804114a5416f0b485af4f9414db1658a935e4ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6471792
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1449343}
NOKEYCHECK=True
GitOrigin-RevId: 7de9cbab6de910af2cc407ad313a00a889c48878