[Editing] Handle deletion when caret is at block boundary
When deleting at the start boundary of a content-editable block/inline
element or at the end boundary of a content-editable block element, if a
hidden select element is present with a non-editable element inside it,
the end of the paragraph was incorrectly determined. This occurred
because the traversal included the invisible element's children, leading
to an incorrect visible position and causing a render crash. The issue
has been resolved by passing the correct candidate node and skipping the
children of invisible nodes to accurately find the end.

Bug: 394637424
Change-Id: I43fe09cd4da9a244afa73955ecc34d89b23bd594
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6471871
Reviewed-by: Sambamurthy Bandaru <sambamurthy.bandaru@microsoft.com>
Reviewed-by: Tanu Jain <tanujain@microsoft.com>
Commit-Queue: Pranav Modi <pranavmodi@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1449828}
NOKEYCHECK=True
GitOrigin-RevId: 0985ff64757159abea627f01b963c1ea2684b5b6