Remove getAllScreensMedia from drive-by web
This CL removes getAllScreensMedia from the drive-by web and
completes the pivot to isolated web apps. The origin trial
and the temporary solution of enabling the origin trial in
isolated contexts by default (while keeping it alive on the
drive-by web) are removed.
This also eliminates the need for the CSP tests (requirement
for the temporary solution), which are removed as well.
All other useful tests are ported to isolated web app tests.
The functionality of these tests remains the same.

Bug: 371971653
Change-Id: I9e0cb231a4e44da7e933d1d7ac32051fb94f5262
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6471575
Commit-Queue: Simon Hangl <simonha@google.com>
Reviewed-by: Elad Alon <eladalon@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1450014}
NOKEYCHECK=True
GitOrigin-RevId: 4bb2a19d57a30a1cea9f57104a5e92c4aba446db