Disallow calc() in named timeline range keyframe selectors
Pure percentage selectors already disallowed calc() expressions. The
named selectors should not be different. It is not obvious from the spec
that this should be the behavior, but having selectors with calc() that
contains relative lengths or sibling-index() would have to be dropped
regardlessly.

Bug: 410883045
Change-Id: I56e0c57f9cdaae9490b47844dad9e06fde09c214
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6476968
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Daniil Sakhapov <sakhapov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1450184}
NOKEYCHECK=True
GitOrigin-RevId: 69e2e4a30da8bedd76f6a1c1953b3d343795824d