[Masonry] Fix BuildGridAxisTracks for row direction
Previously, a masonry container with `masonry-direction: row` would
incorrectly set the direction of its `GridSizingTrackCollection`.
This occurred because the constructor mistakenly interpreted the
direction as the `must_create_baselines` boolean.

This CL fixes the issue by switching the order of the parameters in the
constructor and matching uses in grid to this new expectation.

Bug: 343257585
Change-Id: Iedf5aade27e0597ebfa2dae09849f93468da555d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6476797
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Ethan Jimenez <ethavar@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1450380}
NOKEYCHECK=True
GitOrigin-RevId: ca2feb208a31c56f6cf497546293deb61ab64c16