Implement logic to pause/resume depth sensing
Adds `(pause|resume)depthSensing` methods/behavior from the spec to
allow sites to mitigate the impact of having the depth sensing feature
enabled when they don't need depth (e.g. maybe they only need depth for
a placement scenario and are okay to not receive that data at other
times).

Note to Gardeners: This test relies on WebGL, which is known to not
work on the arm64 WPTs. While I've attempted to update the baselines,
please prefer to simply disable any of the WPTs in this CL on arm64 if
any end up missed.

Bug: 410607163
Change-Id: Id11ac7cf33dd2a310f58701b5c425b170d5446d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6470945
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Reviewed-by: Fred Shih <ffred@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1450623}
NOKEYCHECK=True
GitOrigin-RevId: eb320e714b56c895ce6ce526beba6cbdf4e221ff