VT: Handle empty render pass captures
If the render pass has an empty output rect, we try to create an
empty texture for it, which isn't allowed. This patch makes it
instead reuse the empty (non-existent) render pass path to treat
it as an empty capture, which is what is expected.

R=khushalsagar@chromium.org

Bug: 404478371
Change-Id: Ic94261f0bbcd29dc8f7a62097beb2e754d589522
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6462513
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Commit-Queue: Vladimir Levin <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1450720}
NOKEYCHECK=True
GitOrigin-RevId: 13dadbaa3bf50700f686b7201c744fcd570146b8