[Autofill] Exclude user-agent elements in Autofill
For form controls that live in the user-agent shadow DOM, the
owner/ownee relation does not hold, i.e.,
  base::Contains(GetOwnedFormControls(e.GetOwningFormForAutofill()), e)
isn't true.

As a result, FindFormAndFieldForFormControlElement() hits a
NOTREACHED() when the user interacts with the playback speed of
<video> and <audio> elements, because these are implemented as
<input type=checkbox> in the user-agent DOM.

We cannot fix this in GetOwnedFormControls() because it's intuitively
correct that the function does not descend into user-agent shadow DOMs.

We cannot fix this in GetOwningFormForAutofill() without having a
sentinel return value for user-agent form controls.

Therefore, this CL checks that the form control of interest is not in
the user-agent shadow DOM at the beginning of
FindFormAndFieldForFormControlElement() where we also check if the
form control is connected.

Like for user-agent fields, the owner/ownee relation doesn't hold for
disconnected elements either.

Therefore, this CL introduces the concept of *accessible* elements:
an element is accessible iff it is connected and not in the user-agent
shadow DOM. The CL replaces all occurrences of IsConnected() with
IsAccessible() in Autofill.

Bug: 347059988
Change-Id: I0d877b8a697d707da8dacaf6e8f45820910f275c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6480375
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Commit-Queue: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1450752}
NOKEYCHECK=True
GitOrigin-RevId: fd5a32d9d31c2f54eb275de0ebfa4625ec4fff18