Set up a builder to run Chrome Android Stable
Plan is to use this builder to run upstream WPTs and upload the
results to wpt.fyi.

Do not create a mirror try builder for this one, as it uses a prebuilt
chrome android stable and WPTs from upstream.

This builder runs once everyday. Using the total test time on
android-15-chrome-wpt-fyi-rel as a reference, the number of bots
needed will be 3.

Bug: 402548382
Change-Id: Ib222d53f3cc80ca4def1be2a21b69e7242f752f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6479900
Commit-Queue: Weizhong Xia <weizhong@google.com>
Reviewed-by: Gary Tong <gatong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1450775}
NOKEYCHECK=True
GitOrigin-RevId: 03cbf4014ef7cac610938cd6c0a8a83f536e275c