[perf] Add U+200B to `Character::MaybeBidiRtl`
This function is used to determine whether bidi may be needed
or not. When this function is true (and some other conditions)
we ask ICU to determine if bidi is really needed.

Adding U+200B to this function seems to pay off the cost to
check one additional character, as `InlineItemsBuilder` may
emit U+200B.

Bug: 412935323
Change-Id: I55a311694ac09c9de57bfc4b91499cfb57aa9399
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6483349
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1450933}
NOKEYCHECK=True
GitOrigin-RevId: 7bf3c8531c79e281b31c8839c979460ec86fd197