Simplify calc-size() interpolations more aggressively.
When doing conversions of calc-size() expressions, make sure to call
MaybeSimplifySumOrProductNode(), so that e.g. 25% of (0.5 * size)
becomes (0.125 * size) instead of (0.25 * 0.5 * size). In this process,
we also need to convert kInteger to kNumber during simplification,
or expressions like (8 * 0.25) would not be simplified since 8 and
0.25 are of different units.

Bug: 40118274
Change-Id: I9134d9f0917e7c79504d76b4170315efbf6a6d03
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6483070
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Reviewed-by: Daniil Sakhapov <sakhapov@chromium.org>
Reviewed-by: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1451011}
NOKEYCHECK=True
GitOrigin-RevId: 10da5d6346dccce9c73068756a8ac973188e5106