[source-phase-imports] Update ModuleType to include Wasm
The spec changed the "javascript" module type to "javascript-or-wasm" to
represent both types of modules that come from JavaScript without an
specific type, it later uses the MIME type to disambiguate.

We follow the spec by updating ModuleType::kJavaScript to
ModuleType::kJavaScriptOrWasm and introducing a new enum
`ResolvedModuleType` to be used once the module type is disambiguated
and we need to decide which compiler to call.

Spec: https://html.spec.whatwg.org/multipage/webappapis.html

Bug: 42204365
Change-Id: I35c95ea071c76446879286b00e7a9d373c5d2c04
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6324819
Commit-Queue: Luis Pardo <lpardosixtos@microsoft.com>
Reviewed-by: Dan Clark <daniec@microsoft.com>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1451548}
NOKEYCHECK=True
GitOrigin-RevId: 564725c4918e04e8683274cceeb925f14d3d4438