Make \r and \f abort strings in lazy CSS parsing.
We already had code to make a raw \n inside strings abort
lazy CSS parsing (since they are an edge case that count as
an immediate end quote). However, it turns out that \f and \r
have the same behavior and need to be treated the same;
otherwise, we could find the wrong ending point for blocks
that contained such anomalies.

We use slightly different strategies for the three SIMD paths
we have; SSE2 uses three comparisons and OR (we don't have
anything much better, and this path is only taken by old/slow
machines anyway), AVX2 uses a shuffle-and-compare, and NEON
uses just a single shuffle (which we can do because its
shuffle-index-out-of-bounds behavior is more favorable than
on x86 for this case).

Fixed: 406036299
Change-Id: Id68cc0439ddca8b77ee82740f675b4dd2615b0a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6487751
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1451695}
NOKEYCHECK=True
GitOrigin-RevId: ef179d3aafc37b49395e4f951de23023b386e6f7