Tracing: expose `nodeName` consistently in trace events
This field is useful as on the front end if we only have a nodeId, we
need the trace to be "fresh" (e.g. we are connected to the target) to
resolve info about the Nodes via CDP.

If we have the nodeName in the trace data, we can at least fall back to
that so we can show the user "img.foo" rather than "Node (id: 4)".

Bug: 413284569
Change-Id: Iaa7da0ff71734a52b5fe6e296cd52a0a9b7b2d9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6491448
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1451703}
NOKEYCHECK=True
GitOrigin-RevId: 1c0d122c7ead73f0230730f4d9fc1e5a3a03225d