[line-clamp] Make `text-overflow: ellipsis` work inside `line-clamp`
Since `-webkit-line-clamp` was first implemented in Webkit, using line
clamping would prevent `text-overflow: ellipsis` from working inside
the line clamp container. This seems to have been a side effect of the
initial implementation, and it was kept as it was when the
implementation was redone for LayoutNG, even though that is no longer
a side effect.

Recently, in https://github.com/w3c/csswg-drafts/issues/10823 the
CSSWG resolved that `text-overflow: ellipsis` should apply inside a
line-clamp container. For the last line until clamp, if
`text-overflow: ellipsis` also applies to it, the line clamp ellipsis
is applied first -- however, since we currently only support ellipsis,
rather than a text string, for both the `text-overflow` and line-clamp
ellipses, this would be identical to only applying the line clamp
ellipsis.

This change also removes the `LineClampData::State::kDontTruncate`
state, which was set when there was no possibility of clamping, but
`text-overflow: ellipsis` should still not apply because we were still
inside the line-clamp container. Since `text-overflow: ellipsis` now
should apply in those cases, we remove that state, and use `kDisabled`
instead. We also add an `ignore_line_clamp` flag to
`BlockLineClampData` to know to ignore the line-clamp properties when
performing a relayout without clamping.

Additionally, this change updates some WPT tests that were rendered
incorrect with this resolution.

Bug: 40336192, 407971271
Change-Id: I06e43011a12ba27507b6f661c7b0484d6587cf19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6470150
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Andreu Botella <abotella@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1451710}
NOKEYCHECK=True
GitOrigin-RevId: 0207b9d2a33aa310419238a133133b408dfa6efc