Fix DCHECK error in HTMLCanvasElement::Paint() for OffscreenCanvas frames.
PaintPreviewState was added in Document and was factored into the
"printing" state in WebLocalFrameImpl::CapturePaintPreview, which does
not necessarily update the Document's printing_ state, but rather, the
paint_preview_ state. HTMLCanvasElement::Paint() only accounts for the
Document's printing_ state and not paint_preview_ state, causing the
DCHECK to trigger when OffscreenCanvases are loaded. This CL should fix
this by checking if either the printing_ state or the paint_preview_
state is set to kPrinting or
kPaintingPreview/kPaintingPreviewSkipAcceleratedContent, respectively.
This only happens on chrome-branded builds (i.e.,
is_chrome_branded=true)

Bug: 413405915
Change-Id: I8d4b50e8b954f4b4d640eab94ceab38bee6a3fcc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6489433
Commit-Queue: John Kim <johnykim@google.com>
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1451928}
NOKEYCHECK=True
GitOrigin-RevId: 33d2ddbb92b5544859e7798517205ecdfb569b3e