Ensure quick hover/de-hover before interesttarget delay works
It was previously possible to hover and then quickly de-hover an
interesttarget element, before the show-delay expired, and the
interest would still be shown. Same for a quick focus and blur.
This fixes that bug, plus adds a lot more testing of focus in
combination with the delays. It also splits up the tests using
variants, to avoid timeouts.

Bug: 326681249
Change-Id: Ia43b071723099341e2fbb48934e1eae0e7e09113
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6490548
Auto-Submit: Mason Freed <masonf@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Reviewed-by: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1451945}
NOKEYCHECK=True
GitOrigin-RevId: 2985df153f01dc76ae0ec908096b5bf683c314b1