Change mhtml file type to binary form mhtml files
An internal tool cog ignores .gitattributes and uses GOB storage directly. To fix issues with this tool we need to ensure that the mhtml files indeed have the CRLF encoded in the underlying storage instead of relying on git to convert them on checkout.

Bug: 405964686
Change-Id: Ie06fb5cd7a01e4e1395726489701f37016ab9c5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6490528
Reviewed-by: Nate Chapin <japhet@chromium.org>
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1451995}
NOKEYCHECK=True
GitOrigin-RevId: 77a644c75b022bedaf87e7507a6016631c1da3b2