Fix crash in AddToTopLayer
It is possible, via a convoluted and not very reproducable repro,
to add a ::backdrop pseudo element to something that isn't in the
top layer. See crbug.com/410437529. This CL fixes that by making
sure the element is (still) in the top layer before adding the
backdrop pseudo element to the top layer. It also converts a few
DCHECKs (that hit on the repro) to CHECKs.

Fixed: 410437529
Change-Id: I85d63282a9b7f5161ab8d22e5bfeec013626ad04
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6491968
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Auto-Submit: Mason Freed <masonf@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1452103}
NOKEYCHECK=True
GitOrigin-RevId: 01c7c1b09a3fd84befc0c736dfcce3eac69bea1d