Factor out "is a link click button" logic
Previously this was duplicated between html_anchor_element.cc and
anchor_element_interaction_tracker.cc. This led to an incorrect TODO
comment in the latter which implied we should be checking for
non-default mouse settings, whereas in fact link clicks are always the
left or middle buttons.

By centralizing the logic into MouseEvent::IsLinkClickButton(), it's
more obvious that we're doing the correct thing already, and that TODO
comment can be deleted.

Bug: 409441954
Change-Id: I814c6dc694e27910a45436de55c385dbe0555542
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6491167
Commit-Queue: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1452413}
NOKEYCHECK=True
GitOrigin-RevId: d7873ef2a2e7af2fc4e06e8a29d7ab4a3fdb2529