[gap-decorations] Populate gap intersections for multi-col
This CL adds the logic to populate the `GapGeometry` data structure with
`GapIntersections`.

In short, the way we go about building this is the following: As we
layout each column, we compute the intersections for the gap before that
column, as well as any row intersections. See the comments in
`column_layout_algorithm.h` for a more detailed summary and example.

This CL also moves some testing logic to base_layout_algorithm_test so
that it can be reused for column_layout_algorithm tests.

We also updated the test expectations file for the CSSGapDecorations
flag with some failing tests that we will look at in follow ups.

Bug: 357648037
Change-Id: Ia9ad3aa6e0de7d63b68f1f044d028077ff952c69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6423352
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1453373}
NOKEYCHECK=True
GitOrigin-RevId: 1b61a10980dea9ee0f5dc3f307c482565404e0c9