[Viewport Segments] Per CSS resolution always return one segment.
Per CSS resolution on https://github.com/w3c/csswg-drafts/issues/11957,
always return an array of one segment when a foldable device is used
flat. This also applies to all non-foldable devices which means that
we do not increase the risk of fingerprinting.

This change also match the behavior of MQs (horizontal-viewport-segments for example).
They always match at a value of 1 (see RenderWidgetHostImpl::GetVisualProperties()).

This CL modifies the expected results of tests to reflect the new behavior.

There is no compat issue with this change as the feature is not yet
enabled by default.

Bug: 414430842

Change-Id: Id3cd0fc8ee47a68aadd3ded8d33624d41233f589
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6469455
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Alexis Menard <alexis.menard@intel.com>
Reviewed-by: Alex N. Jose <alexnj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1453468}
NOKEYCHECK=True
GitOrigin-RevId: 9ea51ed248c3b87295c4d33a7510af02acc80e36