[zip] Allow internal symbolic links in unzipper
Allow internal symbolic links to be preserved when extracting a zip
archive. Add an option to control this behavior which defaults to false.

The ability to preserve symbolic links in CRX archives is necessary for
the updater to eliminate reliance on DMG installers for Chrome.

Bug: 411228564
Change-Id: I1ce229a00aeed0947ae8d61f6316facdb3c40387
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6473488
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Reviewed-by: Adenilson Cavalcanti <cavalcantii@chromium.org>
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Commit-Queue: Noah Rose Ledesma <noahrose@google.com>
Cr-Commit-Position: refs/heads/main@{#1453482}
NOKEYCHECK=True
GitOrigin-RevId: d3ff8189237c0298e5a900022d53207bcc419db1