Invalidate filter rect for empty chunks when a filter changes
RasterInvalidator filters invalidations based on the drawable_bounds of
the chunks that are being invalidated. For scenarios where the chunk's
bounds contain all content drawn by the chunk, this is sufficient to
ensure repainting of modified content. But, for chunks where a filter
draws outside of those drawable_bounds, this can lead to changes in a
decomposed filter incorrectly repainting, or even being ignored.

This change addresses two issues that contributed to this bug:

ShouldSkipForRasterInvalidation - was ignoring any chunk that had an
empty drawable bounds (checked via DrawsContent()). To avoid this, this
change overrides !DrawsContent() if the chunk has a reference filter.
This could cause over-invalidation when the chunk is otherwise
effectively_invisible (near-zero opacity).

RasterInvalidator's invalidation rect calculation - PaintChunkInfo is
used by RasterInvalidator to record the rectangle that a chunk may
invalidate on its layer, bounds_in_layer. This bounds is computed by
calling ChunkToLayerMapper::MapVisualRect with the chunk's drawable
bounds. However, that would early out if the passed in bounds was
empty prior to giving reference filters a chance to expand the empty
bounds. Swapping the order, so that the empty rect check happens after
the filter expansion, leads to a correct invalidation rect.

Both of these changes are behind a RuntimeEnabledFeature,
EmptyReferenceFilterInvalidation.

This change also introduces a new unit test,
RasterInvalidatorTest.EffectReferenceFilterChangeOnEmptyChunk, which
validates that chunks with a reference filter and empty drawable bounds
invalidate a rect that includes the filter.

This bug also causes flaky WPT tests. Examples:
css/filter-effects/effect-reference-feimage-001.html
css/filter-effects/effect-reference-feimage-003.html

Bug: 40118302
Change-Id: Iaaca9391c87992c181ba66421e3ac3cb86d1619e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6469305
Commit-Queue: Sam Fortiner <samfort@microsoft.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1453490}
NOKEYCHECK=True
GitOrigin-RevId: f746f478c9148b57d4bc00d5b112fe63379a0d1b