Invalidation tracing: Fix nullptr crash in constructed stylesheet
A constructed stylesheet does not have an ownerNode, so we need to check
for nullptr before dereferencing. This crash would occur when we were
attempting to identify which stylesheet instance we should write to the
log for a given style rule related to an invalidation.

Bug: 337076014
Change-Id: Ic97b5638ed2a41df68e98f5bcc1fd86e2bf27eec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6499127
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1453592}
NOKEYCHECK=True
GitOrigin-RevId: a82d0bc78e8bb27ef92f79e62734460516fc79a8