[LanguageDetector] Consume user activation
This change is a cherry-pick of crrev.com/c/6491262

Consumes user activation if the model hasn't finished downloading yet.
The model is downloaded early on after chrome is launched and is fairly
small so it is unlikely to trigger.

Bug: 412457775
Change-Id: I90230c35cf7f30a618389cacbafb3a361beaa519
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6495556
Commit-Queue: Nathan Memmott <memmott@chromium.org>
Auto-Submit: Christine Hollingsworth <christinesm@chromium.org>
Reviewed-by: Nathan Memmott <memmott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454030}
NOKEYCHECK=True
GitOrigin-RevId: 90a04e5da75ff2543fae2d586d03ef5e4e4f6703