[flex] Rely on LayoutBox cache for flex layout.
This patch does several interconnected things.

The primary purpose of this patch is to make:
 - `width:stretch`
 - `height:stretch`
...stretch to the line cross-size, instead of the container size.

Previously flex-box passed a `LayoutResult` around on `FlexItem` when
the layout-result was valid for the next stage.

This however relied on somewhat complex logic to ensure that we
performed a relayout when the item was `align-self: stretch`.

This logic would need to be expected for the new
`width`/`min-width`/etc keywords.

Instead of this, rely on the LayoutBox cache instead of this
specialized cache. This is marginally slower, but less maintenance.

On SP3 this appears to have a marginal impact of ~0.05%.
This is likely due to the additional work we are performing within
LayoutBox::CachedLayoutResult()

Behind the flag LayoutFlexNewStretch.

An all-in-one (without the flag) version of this patch is at:
https://chromium-review.googlesource.com/c/chromium/src/+/6495824

Change-Id: I6501473f1eadf17c120f48edde0e41f996ee53e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6493160
Reviewed-by: David Grogan <dgrogan@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454073}
NOKEYCHECK=True
GitOrigin-RevId: 6e92c902c58bf2b54d31f4b7e825a0ff85e00627