Refactor: Consolidate multiple video capture callbacks into a struct
This is the second phase of the refactoring effort to group multiple
callback parameters into a single struct. The goal is to simplify
function signatures and make it easier to add new callbacks without
widespread changes.

The VideoCaptureRunState remains separate from the struct, as it is
used at a higher level in the capture pipeline.

Bug: 399692462
Change-Id: I38149de9f1b5a455644a448662460ae66627d944
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6489355
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Sunggook Chue <sunggch@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1454166}
NOKEYCHECK=True
GitOrigin-RevId: 78ce4bcc84111b768039501aabaea496ee396563