Eliminate blocking paint on in-flight image decode
Blink calls HasAlpha() and DecodeFailed() during paint, and that will
block the main thread if the decode is in flight. This CL uses atomics
and a separate base::Lock for the those bits, so they can be queried
safely even while a decode is running.

Bug: chromium:378623805
Change-Id: Ib98f7307178bb2c8828be45ccae5551952bb93fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6481828
Commit-Queue: Stefan Zager <szager@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454204}
NOKEYCHECK=True
GitOrigin-RevId: b266d20110f315eb60b9010d6724e1e6c3be9137